pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "github-ioc-scanner"
version = "1.3.0"
description = "Scan GitHub repositories for Indicators of Compromise (IOCs) in package dependencies across multiple languages"
authors = [{name = "GitHub IOC Scanner Contributors", email = "security@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
keywords = ["security", "supply-chain", "ioc", "github", "scanner", "malware", "typosquatting"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]
dependencies = [
    "httpx>=0.24.0",
    "pyyaml>=6.0",
    "psutil>=5.8.0",
    "pyjwt>=2.8.0",
    "cryptography>=41.0.0",
]

[project.urls]
Homepage = "https://github.com/christianherweg0807/github_package_scanner"
Documentation = "https://github.com/christianherweg0807/github_package_scanner/tree/main/docs"
Repository = "https://github.com/christianherweg0807/github_package_scanner"
Issues = "https://github.com/christianherweg0807/github_package_scanner/issues"
Changelog = "https://github.com/christianherweg0807/github_package_scanner/blob/main/CHANGELOG.md"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
github-ioc-scan = "github_ioc_scanner.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.package-data]
github_ioc_scanner = ["issues/*.py"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true